+++ morpha/morpha/parsers/__init__.py
@@ -2,6 +2,8 @@
 
 from .irecord_parser import IRecordParser
 from .team_preview_record_parser import TeamPreviewRecordParser
+from .battle_log_parser import BattleLogParser
 
-__all__ = ['IRecordParser',
+__all__ = ['BattleLogParser',
+           'IRecordParser',
            'TeamPreviewRecordParser']